Add tariff api folder
